Update entity settings, combine basic editor with normal editor
