+++ src/translations/en.json
@@ -1059,7 +1059,7 @@
           "area": "Set entity area only",
           "area_note": "By default the entities of a device are in the same area as the device. If you change the area of this entity, it will no longer follow the area of the device.",
           "follow_device_area": "Follow device area",
-          "change_device_area": "Change device area",
+          "change_device_area": "change the device area",
           "configure_state": "{integration} options",
           "stream": {
             "preload_stream": "Preload camera stream",
